Upgrades: Domain Managment: support suggesting domain through path
In wp-admin the user was suggested a domain in a tip (or if the domain
was prefilled when buying a bundle) and could follow a link to register
that domain. This PR extends the existing /domains/add path to support
optional suggestion parameter that will prefill the search field, if
present.